feat(core): break api & global split
