+++ package.json
@@ -4,8 +4,8 @@
   "private": true,
   "description": "Typings & utilities for the AE UTBM API & sub-projects",
   "license": "GPL v3",
-  "main": "dist/src/index.js",
-  "types": "src/index.d.ts",
+  "main": "src/index.ts",
+  "types": "types/@ae_utbm/core/index",
   "devDependencies": {
     "@commitlint/cli": "^18.4.3",
     "@commitlint/config-conventional": "^18.4.3",
@@ -20,8 +20,9 @@
     "typescript": "^5.3.2"
   },
   "scripts": {
+    "clean": "node clean.mjs",
     "prepare": "husky install",
-    "prepublish": "tsc",
+    "prepublish": "pnpm clean && tsc",
     "test": "jest --config jest.config.ts"
   }
 }
