--- src/api/constants/genders.ts
@@ -1,2 +1,0 @@
-/* istanbul ignore file */
-export const USER_GENDER = ['OTHER', 'MALE', 'FEMALE'] as const;
