+++ src/constants/genders.ts
@@ -1,0 +1,2 @@
+/* istanbul ignore file */
+export const GENDERS = ['OTHER', 'MALE', 'FEMALE'] as const;
