--- src/index.d.ts
@@ -1,2 +1,0 @@
-export * from './api/@types/index';
-export * from './global/@types/index';
