+++ types/@ae_utbm/api/dto/promotions.d.ts
