+++ types/@ae_utbm/core/index.d.ts
@@ -1,3 +1,7 @@
+export * from '../../../src/index';
+export * from '../api';
+
+// CORE TYPES
 export type aspect_ratio = `${number}:${number}`;
 export type base64 = `data:${string}/${string};base64,${string}`;
 export type email = `${string}@${string}.${string}`;
@@ -6,6 +10,11 @@ export type URL = `${string}://${string}.${string}`;
 
 export type KeysOf<T extends object> = Array<keyof T>;
 
+import { PERMISSIONS, GENDERS } from '../../../src/constants/index';
+export type Permissions = (typeof PERMISSIONS)[number]['name'];
+export type Genders = (typeof GENDERS)[number];
+
+// recursive keys of an object
 export type IsAny<T> = unknown extends T ? ([keyof T] extends [never] ? false : true) : false;
 export type PathImpl<T, Key extends keyof T> = Key extends string
 	? IsAny<T[Key]> extends true
@@ -16,9 +25,7 @@ export type PathImpl<T, Key extends keyof T> = Key extends string
 				| `${Key}.${Exclude<keyof T[Key], keyof unknown[]> & string}`
 		: never
 	: never;
-
 export type PathImpl2<T> = PathImpl<T, keyof T> | keyof T;
-
 export type Path<T> = keyof T extends string
 	? PathImpl2<T> extends infer P
 		? P extends string | keyof T
