+++ types/@ae_utbm/index.d.ts
@@ -1,0 +1,2 @@
+export * from './api';
+export * from './core';
