+++ types/lib.es5.d.ts
@@ -1,5 +1,5 @@
 /**
- * Fix wrong type definitions for toLowerCase and toUpperCase methods in lib.es5.d.ts
+ * Fix "wrong" type definitions for toLowerCase and toUpperCase methods in lib.es5.d.ts
  * @see https://github.com/microsoft/TypeScript/issues/44268
  */
 interface String {
